Add support to cross-grade a subscription on Android
This change introduces basic support to replace an old subscription on
Android. On iOS passing the added parameter will have no effect.

Please note that the current implementation is basic, in that it misses
a way to contextually change the proration mode.

It also uses `setOldSku(string)` rather than `setOldSku(string, string)`
which also requires the old purchaseToken to be provided.

Ref: https://developer.android.com/reference/com/android/billingclient/api/BillingFlowParams.Builder#setoldsku